fix: fix `family_id` field type in `adding family_id field to person table` migration file

+++ database/migrations/2019_10_12_175412_add_family_id_to_person_table.php
@@ -14,7 +14,7 @@ class AddFamilyIdToPersonTable extends Migration
     public function up()
     {
         Schema::table('persons', function (Blueprint $table) {
-            $table->integer('family_id');
+            $table->unsignedBigInteger('family_id');
             $table->foreign('family_id')->references('id')->on('family');
         });
     }
